requirements.txt
anthropic-haystack
hayhooks